new-live: rework content blocks to allow using from code
